experiments/email_reply_bot/requirements.txt
pyyaml>=6.0
google-api-python-client>=2.0
google-auth>=2.0
google-auth-httplib2>=0.1
google-auth-oauthlib>=1.0
anthropic>=0.34
